requirements.txt
requests
pydantic
python-dotenv
tqdm
pandas
gspread
oauth2client
openai